Reorganize components folder for easier navigation
